+++ spriterengine/variable/variablecontainer.cpp
@@ -23,7 +23,7 @@ namespace SpriterEngine
 
 	Variable *VariableContainer::getVariable(int variableIndex)
 	{
-		if (variableIndex<variables.size())
+		if (variableIndex < static_cast<int>(variables.size()))
 		{
 			return variables.at(variableIndex);
 		}
